refactor disk sampler
Refactors disk sampler to move `Statistic` enum out into its own
module
